requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.4
cssselect2==0.8.0
dj-database-url==3.0.0
Django==5.1.7
djangorestframework==3.16.0
et_xmlfile==2.0.0
fonttools==4.58.4
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
numpy==2.2.6
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
pypdf==5.6.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.4.0
uvicorn==0.34.3
webencodings==0.5.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
